refactor: Move <MessageContainer/> and all related components to @chainlit/components
