+++ libs/components/contexts/MessageContext.tsx
@@ -1,0 +1,21 @@
+import { createContext } from 'react';
+
+import { IMessageContext } from '../src/types/messageContext';
+
+const defaultMessageContext = {
+  avatars: [],
+  defaultCollapseContent: false,
+  expandAll: false,
+  hideCot: false,
+  highlightedMessage: null,
+  loading: false,
+  onElementRefClick: undefined,
+  onFeedbackUpdated: undefined,
+  onPlaygroundButtonClick: undefined,
+  showFeedbackButtons: true,
+  uiName: ''
+};
+
+const MessageContext = createContext<IMessageContext>(defaultMessageContext);
+
+export { MessageContext, defaultMessageContext };
