+++ libs/components/hooks/index.ts
@@ -1,3 +1,4 @@
 export { useApi } from './useApi';
 export { useIsDarkMode } from './useIsDarkMode';
 export { useUpload } from './useUpload';
+export { useColors } from './useColors';
