+++ libs/components/src/index.ts
@@ -2,6 +2,7 @@ export * from './auth';
 export * from './buttons';
 export * from './elements';
 export * from './inputs';
+export * from './messages';
 export * from './types';
 
 export { ClipboardCopy } from './ClipboardCopy';
