+++ libs/components/src/messages/components/Author.tsx
@@ -1,25 +1,26 @@
-import { useColorForName } from 'helpers/color';
-import { useRecoilValue } from 'recoil';
+import { useContext } from 'react';
 
 import { Box, Tooltip, Typography } from '@mui/material';
 
-import { AvatarElement, IMessage } from '@chainlit/components';
+import { useColorForName } from '../../../hooks/useColors';
 
-import { avatarState } from 'state/element';
+import { IMessage } from '../../types/message';
 
-import MessageTime from './time';
+import { MessageContext } from '../../../contexts/MessageContext';
+import { AvatarElement } from '../../elements/Avatar';
+import { MessageTime } from './MessageTime';
 
 interface Props {
   message: IMessage;
   show?: boolean;
 }
 
-export const authorBoxWidth = 70;
+export const AUTHOR_BOX_WIDTH = 70;
 
-export default function Author({ message, show }: Props) {
-  const getColorForName = useColorForName();
-  const avatars = useRecoilValue(avatarState);
-  const avatarEl = avatars.find((e) => e.name === message.author);
+const Author = ({ message, show }: Props) => {
+  const context = useContext(MessageContext);
+  const getColorForName = useColorForName(context.uiName);
+  const avatarEl = context.avatars.find((e) => e.name === message.author);
 
   const avatar = show && avatarEl && (
     <AvatarElement element={avatarEl} author={message.author} />
@@ -32,7 +33,7 @@ export default function Author({ message, show }: Props) {
         lineHeight="24px"
         sx={{
           mb: '-5px',
-          width: authorBoxWidth,
+          width: AUTHOR_BOX_WIDTH,
           fontSize: '12px',
           fontWeight: 500,
           color: getColorForName(
@@ -51,7 +52,7 @@ export default function Author({ message, show }: Props) {
 
   return (
     <>
-      <Box width={authorBoxWidth} pr={2}>
+      <Box width={AUTHOR_BOX_WIDTH} pr={2}>
         {display}
         <MessageTime timestamp={message.createdAt} />
       </Box>
@@ -65,4 +66,6 @@ export default function Author({ message, show }: Props) {
       )}
     </>
   );
-}
+};
+
+export { Author };
