+++ libs/components/src/messages/components/ElementRef.tsx
@@ -1,0 +1,31 @@
+import { useContext } from 'react';
+
+import { Link } from '@mui/material';
+
+import { IMessageElement } from '../../types/element';
+
+import { MessageContext } from '../../../contexts/MessageContext';
+
+interface Props {
+  element: IMessageElement;
+}
+
+const ElementRef = ({ element }: Props) => {
+  const { onElementRefClick } = useContext(MessageContext);
+
+  if (element.display === 'inline') {
+    return <span style={{ fontWeight: 700 }}>{element.name}</span>;
+  }
+
+  return (
+    <Link
+      role="link"
+      className="element-link"
+      onClick={() => onElementRefClick && onElementRefClick(element)}
+    >
+      {element.name}
+    </Link>
+  );
+};
+
+export { ElementRef };
