+++ libs/components/src/messages/index.ts
@@ -1,0 +1,1 @@
+export { MessageContainer } from './MessageContainer';
