+++ libs/components/src/types/index.ts
@@ -2,5 +2,6 @@ export * from './action';
 export * from './element';
 export * from './file';
 export * from './Input';
-export * from './messages';
+export * from './message';
+export * from './messageContext';
 export * from './NotificationCount';
